api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
google-generativeai>=0.3.0
Pillow>=10.0.0
